requirements.txt
PyPDF2
pypdf
langchain
chromadb
pinecone
langchain-community
openai
tiktoken
langsmith
streamlit
langchain-openai
langchain-pinecone
ragas
python-docx
black
cohere
langchain_cohere
autogen